[ML] Data frame transform: Fix progress in wizard create step. (#45116)
Fixes a regression where the progress bar in the transform wizard would return to 0% after a transform finishes.

The actual bugfix is just the code in 6033a4c to make sure the code in the wizard uses the same function getTransformProgress() like the transform management page. Also includes a fix for sorting on the transform list table related to progress.

The rest of the diff is a) moving types and utility functions previously only used within the transform management page to the common shared code of data frame transforms so it's available for the wizard without cross-deep-importing (717d894) and b) adding tests with updated mock data (f14e231).
