+++ x-pack/legacy/plugins/ml/public/data_frame/pages/transform_management/components/transform_list/action_delete.test.tsx
@@ -7,10 +7,10 @@
 import { shallow } from 'enzyme';
 import React from 'react';
 
-import { DataFrameTransformListRow } from './common';
+import { DataFrameTransformListRow } from '../../../../common';
 import { DeleteAction } from './action_delete';
 
-import dataFrameTransformListRow from './__mocks__/data_frame_transform_list_row.json';
+import dataFrameTransformListRow from '../../../../common/__mocks__/data_frame_transform_list_row.json';
 
 describe('Data Frame: Transform List Actions <DeleteAction />', () => {
   test('Minimal initialization', () => {
