+++ x-pack/legacy/plugins/ml/public/data_frame/pages/transform_management/components/transform_list/transforms_stats_bar.tsx
@@ -7,7 +7,11 @@
 import React, { FC } from 'react';
 import { i18n } from '@kbn/i18n';
 import { StatsBar, TransformStatsBarStats } from '../../../../../components/stats_bar';
-import { DATA_FRAME_TRANSFORM_STATE, DATA_FRAME_MODE, DataFrameTransformListRow } from './common';
+import {
+  DATA_FRAME_TRANSFORM_STATE,
+  DATA_FRAME_MODE,
+  DataFrameTransformListRow,
+} from '../../../../common';
 
 function createTranformStats(transformsList: DataFrameTransformListRow[]) {
   const transformStats = {
