+++ x-pack/legacy/plugins/ml/public/data_frame/pages/transform_management/services/transform_service/get_transforms.ts
@@ -6,17 +6,14 @@
 
 import { ml } from '../../../../../services/ml_api_service';
 import {
-  DataFrameTransformPivotConfig,
-  refreshTransformList$,
-  REFRESH_TRANSFORM_LIST_STATE,
-} from '../../../../common';
-
-import {
   DataFrameTransformListRow,
   DataFrameTransformStats,
   DATA_FRAME_MODE,
   isDataFrameTransformStats,
-} from '../../components/transform_list/common';
+  DataFrameTransformPivotConfig,
+  refreshTransformList$,
+  REFRESH_TRANSFORM_LIST_STATE,
+} from '../../../../common';
 
 interface GetDataFrameTransformsResponse {
   count: number;
